written in JSX
Remember this isn't a full fix. Filtering should be decoupled from whether the Griddle is open or not if there is not external loading happening. Filtering should also be recursive applying it to the lowest level tbodies and upwards instead of treating only the top level tbody as relevant for filtering.