Stop filter from matching invisible properties
Closes #725. The previous `if (filterable === false)` condition meant
that `filterable={false}` must be explicitly set on a ColumnDefinition
in order for any property returned by row.keySeq() to be ignored. This
resulted in the filter matching invisible row properties such as
griddleKey and any properties in the data not covered by a
ColumnDefinition.

I confirmed this by locally changing the expression at the end of this
filter from:

  return value &&
    value.toString().toLowerCase().indexOf(filterToLower) > -1;

to:

  if (value && value.toString().toLowerCase().indexOf(filterToLower) > -1) {
    console.log('KEY:', key, 'VALUE:', value.toString().toLowerCase());
    return true;
  }
  return false;

and observed the unexpected filter matches on invisible row properties
in the console.

This fixes the problem by updating the condition to exclude griddleKey
and any values not covered by a ColumnDefinition if any
ColumnDefinitions are present. I've added new tests and validated the
filter behavior in the Storybook.

Bear in mind that columns corresponding to other React components may
still match the filter based on generated HTML properties rendered as
part of the value.toString().toLowerCase() expression. At this point the
best course of action may be to recommend that such ColumnDefinitions
are tagged with `filterable={false}` if it's a significant problem for
the user.